Started to build main in MoneyServiceAPP. changes in ExchangeSite

+++ src/MoneyService/Config.java
@@ -66,7 +66,7 @@ public class Config {
 
 	
 	public static MoneyBox fillTheMoneyBox(MoneyBox theBox, Map<String, Currency> currencyMap ) {
-		Map<String, Integer> testMap = new HashMap<String,Integer>(MoneyServiceIO.parseProjectConfig(MoneyServiceIO.readTextFiles(MoneyServiceIO.projectConfigFilename)));
+		Map<String, Double> testMap = new HashMap<String,Double>(MoneyServiceIO.parseProjectConfig(MoneyServiceIO.readTextFiles(MoneyServiceIO.projectConfigFilename)));
 
 		Set<String> keySet = testMap.keySet();
 		for(String k:keySet) {
